+++ Makefile
@@ -7,7 +7,9 @@ SRCS = ${SRCDIR}/json.hpp \
 			 ${SRCDIR}/detail/macro_unscope.hpp \
 			 ${SRCDIR}/detail/meta.hpp \
 			 ${SRCDIR}/detail/exceptions.hpp \
-			 ${SRCDIR}/detail/value_t.hpp
+			 ${SRCDIR}/detail/value_t.hpp \
+			 ${SRCDIR}/detail/conversions/from_json.hpp
+
